requirements.txt
django-rest-framework
Django
celery
django-cors-headers
requests
